Add or set gems (start)
As for items, we may add or set gems, e.g., set level.
In order to start with such feature, we group all gems information in the AllGems dictionary.
We add a first function checkGemEffect which checks that the user-defined gem, given by its effect name, is valid, and returns the corresponding index.